fix: fix task 2 solution

+++ 02-write-file/index.js
@@ -1,29 +1,35 @@
-const fs = require('fs').promises;
-const path = require('path');
+import fs from 'fs/promises';
+import path from 'path';
 
 const stdout = process.stdout;
 const stdin = process.stdin;
 
-fs.writeFile(
-  path.join(__dirname, 'text.txt'),
-  '',
-  (err) => {
-    if (err) throw err;
-  });
+async function isFileExist (file) {
+  try {
+    await fs.access(file);
+    return true;
+  } catch (err) {
+    return false;
+  }
+}
 
 stdout.write('Please, enter your message\n');
 
-stdin.on('data', (data) => {
+stdin.on('data', async (data) => {
   if (data.includes('exit')) {
     process.exit(); 
   }
-  fs.appendFile(
-    path.join(__dirname, 'text.txt'),
-    data,
-    err => {
-      if (err) throw err;
-    });
+  if (await !isFileExist('02-write-file/text.txt')) {
+    await fs.writeFile(
+      path.join('02-write-file/', 'text.txt'),
+      data);
+  } else {
+    await fs.appendFile(
+      path.join('02-write-file/', 'text.txt'),
+      data);
+  }
 });
+
 process.on('SIGINT', function() {
   process.exit();
 });
